requirements.txt
# BlinkTracking V2.0 Requirements

# Core
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0

# Data handling
pandas>=2.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0

# Optional: GPU support
# onnxruntime-gpu>=1.15.0

# Development
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
